requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
docopt==0.6.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
lxml==5.3.0
pipreqs==0.4.13
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
reqs==0.0.2
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.45.1
yarg==0.1.10
